// src/app/model/class/project.ts
export interface Project {
    clientProjectId: number
    projectName: string
    startDate: string
    expectedEndDate: string  
    leadByEmpId: number
    completedDate: string
    contactPersion: string
    contactPersonContactNo: string
    totalEmpWorking: number
    projectCost: number
    projectDetail: string
    contactPersonEmailId: string
    clientId: number
  }